Ended c++ porting

+++ CPorting/Tfce.h
@@ -1,5 +1,5 @@
 #ifndef TFCE_H
 #define TFCE_H
 #include <float.h>
-double * tfce_score(double * map, int dim_x, int dim_y, int dim_z, double E, double H, double dh);
+float * tfce_score(float * map, int dim_x, int dim_y, int dim_z, float E, float H, float dh);
 #endif //TFCE_H
