+++ CPorting/Utilities.cpp
@@ -5,7 +5,7 @@
 #include <string.h>
 
 //Finds minimum, maximum and range of a float vector  
-void findMinMax(double *map, int n, double *min, double *max, double * range) {
+void findMinMax(float *map, int n, float *min, float *max, float * range) {
 	(*min) = FLT_MAX;
 	(*max) = FLT_MIN;
 	for (int i = 0; i < n; ++i) {
@@ -18,20 +18,20 @@ void findMinMax(double *map, int n, double *min, double *max, double * range) {
 }
 
 //performs confrontation between and b given a single char operation
-int lessThan(double a, double b){
+int lessThan(float a, float b){
         return a < b;
 }
 
-int equalTo(double a, double b){
+int equalTo(float a, float b){
         return a == b;
 }
 
-int moreThan(double a, double b){
+int moreThan(float a, float b){
         return a > b;
 }
 
 //Gets a binary vector in matlab style, given a certain operation
-int * getBinaryVector(double * map, int n, int (*confront)(double, double), double value, int * numOfElementsMatching){
+int * getBinaryVector(float * map, int n, int (*confront)(float, float), float value, int * numOfElementsMatching){
 	int * binaryVector = new int [n];//(int *) calloc(sizeof(int), n);
 	int i;
 
@@ -47,8 +47,8 @@ int * getBinaryVector(double * map, int n, int (*confront)(double, double), doub
 }
 
 //Given a binary vector and a real vector, fills a new vector in the indices where binary == 1
-double * fromBinaryToRealVector(double * map, int n, int * binaryVector){
-	double * toReturn = new double[n];//(double *) calloc(sizeof(double), n);
+float * fromBinaryToRealVector(float * map, int n, int * binaryVector){
+	float * toReturn = new float[n];//(float *) calloc(sizeof(float), n);
 	int i;
 
 	for (i = 0; i < n; ++i) {
@@ -61,9 +61,9 @@ double * fromBinaryToRealVector(double * map, int n, int * binaryVector){
 	return toReturn;
 }
 
-//creates a n-element vector of doubles filled with 0
-double * fill0(int n){
-	double * toReturn = new double[n];//(double *) calloc(sizeof(double), n);
+//creates a n-element vector of floats filled with 0
+float * fill0(int n){
+	float * toReturn = new float[n];//(float *) calloc(sizeof(float), n);
 	int i;
 	for (i = 0; i < n; ++i) {
 		toReturn[i] = 0;
@@ -71,22 +71,22 @@ double * fill0(int n){
 	return toReturn;
 }
 
-double multiply(double a, double b){
+float multiply(float a, float b){
 	return a * b;
 }
 
-double elevate(double a, double b){
+float elevate(float a, float b){
 	return pow(a,b);
 }
 
-void apply_function(double * vector, int n, double (* operation) (double a, double b), double argument){
+void apply_function(float * vector, int n, float (* operation) (float a, float b), float argument){
 	int i;
 	for (i = 0; i < n; ++i) {
 	    vector[i] = operation(vector[i], argument);
 	}
 }
 
-void abs_vector(double * vector, int n){
+void abs_vector(float * vector, int n){
 	int i;
 	for (i = 0; i < n; ++i) {
 		if(vector[i] < 0)
@@ -108,21 +108,21 @@ void coordinatesFromLinearIndex(int index, int max_x, int max_y, int * x, int * 
 	(*z) = index;
 }
 
-//copies an int vector into a vector of doubles
-double * copyAndConvertIntVector(int * vector, int n){
-	double * toReturn = new double[n];//(double *) calloc(sizeof(double), n);
+//copies an int vector into a vector of floats
+float * copyAndConvertIntVector(int * vector, int n){
+	float * toReturn = new float[n];//(float *) calloc(sizeof(float), n);
 	int i;
 	for (i = 0; i < n; ++i) {
-		toReturn[i] = (double)vector[i];
+		toReturn[i] = (float)vector[i];
 	}
 	return toReturn;
 }
 
-double * readMatFromFile(FILE *fp, int * size, int * x, int * y, int * z){
+float * readMatFromFile(FILE *fp, int * size, int * x, int * y, int * z){
 	char * line = NULL;
 	char * oldline = NULL;
 	char *token;
-	double *mat;
+	float *mat;
 	size_t len = 0;
 	ssize_t read;
 
@@ -145,12 +145,12 @@ double * readMatFromFile(FILE *fp, int * size, int * x, int * y, int * z){
 				num_of_val++;
 			}
 			//printf("Num of val %d\n", num_of_val);
-			mat  = (double *) calloc(num_of_val, sizeof(double));
+			mat  = (float *) calloc(num_of_val, sizeof(float));
 
 			num_of_val = 0;
 			
 			for (token = strtok(oldline, " "); token != NULL; token = strtok(NULL, " ")) {
-				mat[num_of_val] = (double)atof(token);
+				mat[num_of_val] = (float)atof(token);
 				//printf("%f\n",mat[num_of_val]);
 				num_of_val++;
 			}
@@ -187,9 +187,9 @@ static int rand_int(int n) {
 	return rnd % n;
 }
 
-void * shuffle(double *array, int n) {
+void * shuffle(float *array, int n) {
 	int i, j;
-	double  tmp;
+	float  tmp;
 
 	for (i = n - 1; i > 0; i--) {
 		j = rand_int(i + 1);
@@ -202,9 +202,9 @@ void * shuffle(double *array, int n) {
 
 void printToFile(const char * filename, void * map, int n, int isInteger){
 	int i;
-	double * clustered_map;
+	float * clustered_map;
 	int * other_map =  (int *) map;
-	clustered_map = (double *) map;
+	clustered_map = (float *) map;
 	FILE * log = fopen(filename, "w");
 	if(!isInteger){
 		for(i=0; i < n; i++){
